Refactor server related code in its own module.
