+++ rust/operator-binary/src/connect/mod.rs
@@ -1,2 +1,4 @@
+mod common;
 pub mod controller;
 pub mod crd;
+pub mod server;
